Resampling Capabilities
I added resampling capabilities to the DataFrame class. However it currently there's no way to specify the interval you want the data resampled in. It only works with weeks. Resampling with other time frames isn't too difficult to add but its not something I have for (at least in the near future).